+++ packages/opencensus-core/test/test-stats.ts
@@ -15,7 +15,7 @@
  */
 
 import * as assert from 'assert';
-import {BaseView, Stats, StatsEventListener} from '../src';
+import {BaseView, globalStats, StatsEventListener} from '../src';
 import {AggregationType, LastValueData, Measure, Measurement, MeasureType, MeasureUnit, View} from '../src/stats/types';
 
 class TestExporter implements StatsEventListener {
@@ -41,10 +41,8 @@ class TestExporter implements StatsEventListener {
 }
 
 describe('Stats', () => {
-  let stats: Stats;
-
-  beforeEach(() => {
-    stats = new Stats();
+  afterEach(() => {
+    globalStats.clear();
   });
 
   const viewName = 'testViewName';
@@ -56,8 +54,8 @@ describe('Stats', () => {
 
   describe('createMeasureDouble()', () => {
     it('should create a measure of type double', () => {
-      const measureDouble =
-          stats.createMeasureDouble(measureName, measureUnit, description);
+      const measureDouble = globalStats.createMeasureDouble(
+          measureName, measureUnit, description);
       assert.strictEqual(measureDouble.type, MeasureType.DOUBLE);
       assert.strictEqual(measureDouble.name, measureName);
       assert.strictEqual(measureDouble.unit, measureUnit);
@@ -68,7 +66,7 @@ describe('Stats', () => {
   describe('createMeasureInt64()', () => {
     it('should create a measure of type int64', () => {
       const measureDouble =
-          stats.createMeasureInt64(measureName, measureUnit, description);
+          globalStats.createMeasureInt64(measureName, measureUnit, description);
       assert.strictEqual(measureDouble.type, MeasureType.INT64);
       assert.strictEqual(measureDouble.name, measureName);
       assert.strictEqual(measureDouble.unit, measureUnit);
@@ -84,16 +82,21 @@ describe('Stats', () => {
     let measure: Measure;
 
     before(() => {
-      measure = stats.createMeasureInt64(measureName, measureUnit);
+      measure = globalStats.createMeasureInt64(measureName, measureUnit);
+    });
+
+    after(() => {
+      globalStats.clear();
     });
 
     for (const aggregationType of aggregationTypes) {
       it(`should create a view with ${aggregationType} aggregation`, () => {
         const bucketBoundaries =
             AggregationType.DISTRIBUTION ? [1, 2, 3] : null;
-        const view = stats.createView(
+        const view = globalStats.createView(
             viewName, measure, aggregationType, tagKeys, description,
             bucketBoundaries);
+        globalStats.registerView(view);
 
         assert.strictEqual(view.name, viewName);
         assert.strictEqual(view.measure, measure);
@@ -106,7 +109,7 @@ describe('Stats', () => {
 
     it('should not create a view with distribution aggregation when no bucket boundaries were given',
        () => {
-         assert.throws(stats.createView, 'No bucketBoundaries specified');
+         assert.throws(globalStats.createView, 'No bucketBoundaries specified');
        });
   });
 
@@ -115,18 +118,19 @@ describe('Stats', () => {
     const testExporter = new TestExporter();
 
     before(() => {
-      measure = stats.createMeasureInt64(measureName, measureUnit);
+      testExporter.clean();
+      measure = globalStats.createMeasureInt64(measureName, measureUnit);
     });
 
     it('should register a view', () => {
-      stats.registerExporter(testExporter);
+      globalStats.registerExporter(testExporter);
       const view = new BaseView(
           viewName, measure, AggregationType.LAST_VALUE, tagKeys, description);
 
       assert.ok(!view.registered);
       assert.strictEqual(testExporter.registeredViews.length, 0);
 
-      stats.registerView(view);
+      globalStats.registerView(view);
 
       assert.ok(view.registered);
       assert.strictEqual(testExporter.registeredViews.length, 1);
@@ -139,20 +143,25 @@ describe('Stats', () => {
     const testExporter = new TestExporter();
     let aggregationData: LastValueData;
     before(() => {
-      measure = stats.createMeasureInt64(measureName, measureUnit);
+      measure = globalStats.createMeasureInt64(measureName, measureUnit);
     });
 
     beforeEach(() => {
       testExporter.clean();
-      stats.registerExporter(testExporter);
-      stats.createView(
+      globalStats.registerExporter(testExporter);
+      const view = globalStats.createView(
           viewName, measure, AggregationType.LAST_VALUE, tagKeys, description);
+      globalStats.registerView(view);
+    });
+
+    afterEach(() => {
+      globalStats.clear();
     });
 
     it('should record a single measurement', () => {
       const measurement = {measure, tags, value: 1};
       assert.strictEqual(testExporter.recordedMeasurements.length, 0);
-      stats.record(measurement);
+      globalStats.record(measurement);
       assert.strictEqual(testExporter.recordedMeasurements.length, 1);
       assert.deepEqual(testExporter.recordedMeasurements[0], measurement);
       aggregationData =
@@ -161,9 +170,9 @@ describe('Stats', () => {
     });
 
     it('should not record a single negative measurement', () => {
-      stats.registerExporter(testExporter);
+      globalStats.registerExporter(testExporter);
       const measurement = {measure, tags, value: -1};
-      stats.record(measurement);
+      globalStats.record(measurement);
       assert.strictEqual(testExporter.recordedMeasurements.length, 0);
     });
 
@@ -171,7 +180,7 @@ describe('Stats', () => {
       const measurement1 = {measure, tags, value: 1};
       const measurement2 = {measure, tags, value: 1};
       assert.strictEqual(testExporter.recordedMeasurements.length, 0);
-      stats.record(measurement1, measurement2);
+      globalStats.record(measurement1, measurement2);
       assert.strictEqual(testExporter.recordedMeasurements.length, 2);
       assert.deepEqual(testExporter.recordedMeasurements[0], measurement1);
       assert.deepEqual(testExporter.recordedMeasurements[1], measurement2);
@@ -186,7 +195,7 @@ describe('Stats', () => {
            {measure, tags, value: 1}, {measure, tags, value: -1},
            {measure, tags, value: 1}
          ];
-         stats.record(...measurments);
+         globalStats.record(...measurments);
          assert.equal(testExporter.recordedMeasurements.length, 0);
        });
   });